[typescript] 1000 lines from domui.js done.
