+++ to.etc.domui/src/main/resources/resources/ts/domui.jquery.d.ts
@@ -2,6 +2,7 @@ interface JQuery {
 	webui(xml: any) : void;
 	disableSelection(): void;
 	enableSelection():  void;
+	error(fn: Function):  void;
 }
 
 interface JQueryStatic {
